requirements.txt
Click==7.0
dnspython==1.16.0
email-validator==1.0.4
fastapi==0.65.2
h11==0.8.1
idna==2.8
pydantic==1.8.2
starlette==0.14.2
uvicorn==0.14.0
websockets==7.0
pytest
py2neo
neo4j
python-dotenv
coverage
flake8
mypy
tox-gh-actions
